requirements.txt
SQLAlchemy~=2.0.38
pydantic~=2.11.0b1
pytz~=2025.1
pytest~=8.3.5
python-dotenv~=1.0.1
APScheduler~=4.0.0a5